Feat/#123: useOpenModal 커스텀훅 분리
