+++ src/pages/CommunityDetail/hooks/index.ts
@@ -1,4 +1,5 @@
 import useImageUpload from './useImageUpload';
 import useTextInput from './useTextInput';
+import useToastOpen from './useToastOpen';
 
-export { useTextInput, useImageUpload };
+export { useTextInput, useImageUpload, useToastOpen };
