+++ src/pages/CommunityDetail/hooks/useImageUpload.ts
@@ -1,10 +1,9 @@
 import { useState } from 'react';
 
-const useImageUpload = () => {
+const useImageUpload = (handleModalOpen: () => void) => {
   const [imageSelected, setImageSelected] = useState<boolean>(false);
   const [imageName, setImageName] = useState<string>('');
   const [imageFile, setImageFile] = useState<File | null>(null);
-  const [isToastOpen, setIsToastOpen] = useState<boolean>(false);
   const [toastType, setToastType] = useState<'sizeErr' | 'ResubmitErr' | null>(null);
 
   const handleImageChange = (isSelected: boolean, fileName: string, file: File | null) => {
@@ -33,23 +32,16 @@ const useImageUpload = () => {
     setToastType('ResubmitErr');
   };
 
-  const handleModalOpen = () => {
-    setIsToastOpen(true);
-    setTimeout(() => setIsToastOpen(false), 300000);
-  };
-
   return {
     toastType,
     imageSelected,
     imageName,
     imageFile,
-    isToastOpen,
     handleImageChange,
     handleImgReSubmit,
     handleImageRemove,
     handleSizeError,
     handleResubmitError,
-    handleModalOpen,
   };
 };
